Use existing events.
Move functionality from NotificationsViewModel into
HomeworkspaceViewModel. Remove the Notifications class.
