+++ src/DynamoCore/Models/DynamoModelEventArgs.cs
@@ -127,13 +127,6 @@ namespace Dynamo.Models
     {
         private readonly IOption<Exception> error;
 
-        public EvaluationCompletedEventArgs(bool evaluationTookPlace, Exception errorMsg = null)
-        {
-            EvaluationTookPlace = evaluationTookPlace;
-
-            error = errorMsg != null ? Option.Some(errorMsg) : Option.None<Exception>();
-        }
-
         public bool EvaluationTookPlace { get; private set; }
 
         public bool EvaluationSucceeded
@@ -153,6 +146,13 @@ namespace Dynamo.Models
                             "Evaluation success, no error message recorded.");
                     });
             }
+        }
+
+        public EvaluationCompletedEventArgs(bool evaluationTookPlace, Exception errorMsg = null)
+        {
+            EvaluationTookPlace = evaluationTookPlace;
+
+            error = errorMsg != null ? Option.Some(errorMsg) : Option.None<Exception>();
         }
     }
