+++ src/DynamoCoreWpf/DynamoCoreWpf.csproj
@@ -201,7 +201,6 @@
     <Compile Include="Utilities\ResourceUtilities.cs" />
     <Compile Include="ViewModels\Core\DynamoViewModelBranding.cs" />
     <Compile Include="ViewModels\Core\HomeWorkspaceViewModel.cs" />
-    <Compile Include="ViewModels\NotificationsViewModel.cs" />
     <Compile Include="ViewModels\RunSettingsViewModel.cs" />
     <Compile Include="ViewModels\Search\BrowserInternalElementViewModel.cs" />
     <Compile Include="ViewModels\Search\BrowserItemViewModel.cs" />
