+++ test/DynamoCoreUITests/RunSettingsTests.cs
@@ -2,6 +2,7 @@
 using System.Windows;
 using System.Windows.Input;
 using System.Windows.Media;
+
 using System.Linq;
 
 using SystemTestServices;
